requirements.txt
langchain-openai>=0.0.5
langchain>=0.1.0
langgraph>=0.0.10
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.2.0
numpy>=1.26.0
